@@ -1,29 +1,25 @@
 name: flutter_app
 description: A new Flutter application.
 
-
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.0+1
+version: 1.1.0
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter_lints: ^2.0.1
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.2
-
   url_launcher: ^6.0.3
-
   context_menus:
     path: ../
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
 
 # The following section is specific to Flutter.
 flutter:
