@@ -1,15 +1,16 @@
 name: context_menus
 description: A package to show context menus on right-click or long-press.
-version: 1.0.2
+version: 2.0.0+1
 homepage: https://github.com/gskinnerTeam/flutter-context-menu
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  url_launcher: ^6.2.6
 
 dev_dependencies:
   flutter_test:
