@@ -1,14 +1,22 @@
 import 'package:flutter/material.dart';
 
-typedef Widget ContextMenuButtonBuilder(BuildContext context, ContextMenuButtonConfig config,
-    [ContextMenuButtonStyle? style]);
+typedef Widget ContextMenuButtonBuilder(
+  BuildContext context,
+  ContextMenuButtonConfig config, [
+  ContextMenuButtonStyle? style,
+]);
 
 /// The default ContextMenu button. To provide your own, override [ContextMenuOverlay] buttonBuilder.
+/// TODO: This should use FocusableControlBuilder?
 class ContextMenuButton extends StatefulWidget {
+  const ContextMenuButton(
+    this.config, {
+    this.style,
+    super.key,
+  });
+
   final ContextMenuButtonConfig config;
   final ContextMenuButtonStyle? style;
-
-  const ContextMenuButton(this.config, {Key? key, this.style}) : super(key: key);
 
   @override
   _ContextMenuButtonState createState() => _ContextMenuButtonState();
@@ -16,24 +24,30 @@ class ContextMenuButton extends StatefulWidget {
 
 class _ContextMenuButtonState extends State<ContextMenuButton> {
   bool _isMouseOver = false;
-  set isMouseOver(bool isMouseOver) => setState(() => _isMouseOver = isMouseOver);
+
+  set isMouseOver(bool isMouseOver) {
+    setState(() => _isMouseOver = isMouseOver);
+  }
+
   ContextMenuButtonConfig get config => widget.config;
 
   @override
   Widget build(BuildContext context) {
     bool isDisabled = widget.config.onPressed == null;
     bool showMouseOver = _isMouseOver && !isDisabled;
-    Color defaultTextColor = Theme.of(context).brightness == Brightness.dark ? Colors.white : Colors.black;
+
     final ThemeData theme = Theme.of(context);
-    TextStyle? defaultTextStyle = theme.textTheme.bodyText1?.copyWith(color: theme.colorScheme.onSecondary);
-    TextStyle? shortcutTextStyle = theme.textTheme.bodyText2?.copyWith(color: theme.colorScheme.onSecondary);
+    Color defaultTextColor = theme.brightness == Brightness.dark ? Colors.white : Colors.black;
+    TextStyle? defaultTextStyle = theme.textTheme.bodyMedium?.copyWith(color: theme.colorScheme.onSecondary);
+    TextStyle? shortcutTextStyle = theme.textTheme.bodySmall?.copyWith(color: theme.colorScheme.onSecondary);
+
     ContextMenuButtonStyle style = ContextMenuButtonStyle(
       textStyle: widget.style?.textStyle ?? defaultTextStyle,
       shortcutTextStyle: widget.style?.shortcutTextStyle ?? shortcutTextStyle,
       fgColor: widget.style?.fgColor ?? defaultTextColor,
       bgColor: widget.style?.bgColor ?? Colors.transparent,
-      hoverBgColor: widget.style?.hoverBgColor ?? Theme.of(context).backgroundColor.withOpacity(.2),
-      hoverFgColor: widget.style?.hoverFgColor ?? Theme.of(context).colorScheme.secondary,
+      hoverBgColor: widget.style?.hoverBgColor ?? theme.colorScheme.primaryContainer,
+      hoverFgColor: widget.style?.hoverFgColor ?? theme.colorScheme.secondary,
       padding: widget.style?.padding ?? EdgeInsets.all(6),
     );
 
@@ -60,13 +74,20 @@ class _ContextMenuButtonState extends State<ContextMenuButton> {
                 /// Optional Icon
                 if (config.icon != null) ...[
                   SizedBox(
-                      width: 16, height: 16, child: (showMouseOver) ? config.iconHover ?? config.icon! : config.icon!),
+                    width: 16,
+                    height: 16,
+                    child: (showMouseOver) ? config.iconHover ?? config.icon! : config.icon!,
+                  ),
                   SizedBox(width: 16)
                 ],
 
                 /// Main Label
-                Text(config.label,
-                    style: style.textStyle!.copyWith(color: showMouseOver ? style.hoverFgColor : style.fgColor)),
+                Text(
+                  config.label,
+                  style: style.textStyle!.copyWith(
+                    color: showMouseOver ? style.hoverFgColor : style.fgColor,
+                  ),
+                ),
                 Spacer(),
 
                 /// Shortcut Label
@@ -90,15 +111,17 @@ class _ContextMenuButtonState extends State<ContextMenuButton> {
 }
 
 class ContextMenuButtonStyle {
-  const ContextMenuButtonStyle(
-      {this.fgColor,
-      this.bgColor,
-      this.hoverFgColor,
-      this.hoverBgColor,
-      this.padding,
-      this.textStyle,
-      this.shortcutTextStyle,
-      this.disabledOpacity = .7});
+  const ContextMenuButtonStyle({
+    this.fgColor,
+    this.bgColor,
+    this.hoverFgColor,
+    this.hoverBgColor,
+    this.padding,
+    this.textStyle,
+    this.shortcutTextStyle,
+    this.disabledOpacity = 0.7,
+  });
+
   final Color? fgColor;
   final Color? bgColor;
   final Color? hoverFgColor;
@@ -108,15 +131,16 @@ class ContextMenuButtonStyle {
   final TextStyle? shortcutTextStyle;
   final double disabledOpacity;
 
-  ContextMenuButtonStyle copyWith(
-      {Color? fgColor,
-      Color? bgColor,
-      Color? hoverFgColor,
-      Color? hoverBgColor,
-      EdgeInsets? padding,
-      TextStyle? textStyle,
-      TextStyle? shortcutTextStyle,
-      double disabledOpacity = .7}) {
+  ContextMenuButtonStyle copyWith({
+    Color? fgColor,
+    Color? bgColor,
+    Color? hoverFgColor,
+    Color? hoverBgColor,
+    EdgeInsets? padding,
+    TextStyle? textStyle,
+    TextStyle? shortcutTextStyle,
+    double disabledOpacity = 0.7,
+  }) {
     return ContextMenuButtonStyle(
       fgColor: fgColor ?? this.fgColor,
       bgColor: bgColor ?? this.bgColor,
@@ -130,11 +154,17 @@ class ContextMenuButtonStyle {
 }
 
 class ContextMenuButtonConfig {
+  ContextMenuButtonConfig(
+    this.label, {
+    required this.onPressed,
+    this.shortcutLabel,
+    this.icon,
+    this.iconHover,
+  });
+
   final String label;
   final String? shortcutLabel;
   final VoidCallback? onPressed;
   final Widget? icon;
   final Widget? iconHover;
-
-  ContextMenuButtonConfig(this.label, {required this.onPressed, this.shortcutLabel, this.icon, this.iconHover});
 }
